Переносит функции пропуска кадров и устранения дребезга из main.js в utils.js
