+++ js/error-message.js
@@ -1,5 +1,4 @@
 const REMOVE_MESSAGE_TIMEOUT = 5000;
-
 const errorLoadDataTempleate = document.querySelector('#data-error').content;
 const body = document.body;
