+++ js/utils.js
@@ -31,4 +31,27 @@ const numDecline = (num, nominative, genitiveSingular, genitivePlural) => {
   return num % 10 === 1 ? nominative : genitiveSingular;
 };
 
-export{getRandomInteger, retrieveIndices, createRangeOfNumbers, isEscapeKey, numDecline};
+function debounce (callback, timeoutDelay = 500) {
+
+  let timeoutId;
+  return (...rest) => {
+    clearTimeout(timeoutId);
+
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+}
+
+function throttle (callback, delayBetweenFrames) {
+  let lastTime = 0;
+
+  return (...rest) => {
+    const now = new Date();
+
+    if (now - lastTime >= delayBetweenFrames) {
+      callback.apply(this, rest);
+      lastTime = now;
+    }
+  };
+}
+
+export{getRandomInteger, retrieveIndices, createRangeOfNumbers, isEscapeKey, numDecline, debounce, throttle};
